Updated MLP example
