+++ README.md
@@ -6,15 +6,15 @@
     <th>Linux</th>
   </tr>
   <tr>
-    <td><img src="https://camo.githubusercontent.com/5ce01fbc787bc0f3ff51358ff9a00ad4ada15b08e9c987e9760328fb9667435d/68747470733a2f2f7472617669732d6d61747269782d6261646765732e6865726f6b756170702e636f6d2f7265706f732f7072616268756f6d6b61722f7079746f7263682d6370702f6272616e636865732f6d61737465722f37"></td>
-    <td><img src="https://camo.githubusercontent.com/5ce01fbc787bc0f3ff51358ff9a00ad4ada15b08e9c987e9760328fb9667435d/68747470733a2f2f7472617669732d6d61747269782d6261646765732e6865726f6b756170702e636f6d2f7265706f732f7072616268756f6d6b61722f7079746f7263682d6370702f6272616e636865732f6d61737465722f37"></td>
+    <td><img src="https://img.shields.io/badge/build-failing-red"></td>
+    <td><img src="https://img.shields.io/badge/build-passing-green"></td>
   </tr>
 </table>
 
-1. [Linear Regression](./libtorch-examples/include/linear_regression.hpp)
+1. [Linear Regression](./libtorch-examples/linear-regression)
 
-2. [Logistic Regression](./libtorch-examples/include/logistic_regression.hpp)
+2. [Logistic Regression](./libtorch-examples/logistic-regression)
 
-3. [Multi-Layered Perceptron](./libtorch-examples/include/multi_layered_perceptron.hpp)
+3. [Multi-Layered Perceptron](./libtorch-examples/multi-layered-perceptron)
 
 4. [Convnet Classifier](./libtorch-examples/include/convnet_classifier.hpp)
