+++ libtorch-examples/multi-layered-perceptron/CMakeLists.txt
@@ -5,6 +5,6 @@ set(CMAKE_CXX_STANDARD 14)
 set(EXECUTABLE_NAME multi-layered-perceptron)
 
 add_executable(${EXECUTABLE_NAME})
-target_sources(${EXECUTABLE_NAME} PRIVATE main.cpp include/mlp.hpp)
+target_sources(${EXECUTABLE_NAME} PRIVATE main.cpp include/mlp.hpp include/mlp_classifier.hpp)
 target_include_directories(${EXECUTABLE_NAME} PRIVATE include)
 target_link_libraries(${EXECUTABLE_NAME} ${TORCH_LIBRARIES})
