+++ libtorch-examples/multi-layered-perceptron/main.cpp
@@ -1,8 +1,10 @@
-#include "mlp.hpp"
+#include "mlp_classifier.hpp"
 
 int main() {
     MLPClassifier classifier;
     classifier.compile(784, 500, 0.001);
-    classifier.train("../../../data/mnist", 100, 5, "../../checkpoints");
+    classifier.train(
+            "../../../data/mnist", 100,
+            5, "../../checkpoints");
     classifier.evaluate("../../../data/mnist", 100);
 }
